requirements.txt
pytest==7.4.3
selenium==4.15.2
allure-pytest==2.13.2
requests==2.31.0

